+++ packages/db/dev/vitest.config.ts
@@ -1,0 +1,14 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  esbuild: {
+    target: 'esnext',
+  },
+  test: {
+    include: ['src/**/*.test.ts'],
+    testTimeout: 30_000,
+    typecheck: {
+      enabled: true,
+    },
+  },
+})