+++ packages/dev-utils/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [4.3.3](https://github.com/netlify/primitives/compare/dev-utils-v4.3.2...dev-utils-v4.3.3) (2025-12-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @netlify/types bumped from 2.2.0 to 2.3.0
+
+## [4.3.2](https://github.com/netlify/primitives/compare/dev-utils-v4.3.1...dev-utils-v4.3.2) (2025-11-14)
+
+
+### Bug Fixes
+
+* **deps:** update dependency uuid to v13 ([#539](https://github.com/netlify/primitives/issues/539)) ([aedb279](https://github.com/netlify/primitives/commit/aedb279a24ad41d1db5db98d1e0e976e4a2db98a))
+
+## [4.3.1](https://github.com/netlify/primitives/compare/dev-utils-v4.3.0...dev-utils-v4.3.1) (2025-11-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @netlify/types bumped from 2.1.0 to 2.2.0
+
 ## [4.3.0](https://github.com/netlify/primitives/compare/dev-utils-v4.2.0...dev-utils-v4.3.0) (2025-10-10)